# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "coding-agent-api",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks/{task_id}/cleanup/branch": {
      "post": {
        "operationId": "delete_branch",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeliveryDeleteBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/cleanup/worktree": {
      "post": {
        "operationId": "remove_worktree",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeliveryRemoveWorktreeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}/merge": {
      "post": {
        "operationId": "accept_merge",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeliveryMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/merge/preflight": {
      "post": {
        "operationId": "preflight",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeliveryPreflightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "request_id",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
...
      "DeliveryCommandResponse": {
        "type": "object",
        "required": [
          "receipt",
          "operation"
        ],
        "properties": {
          "operation": {
            "$ref": "#/components/schemas/DeliveryOperationDto"
          },
          "receipt": {
            "$ref": "#/components/schemas/DeliveryReceiptDispositionDto"
          }
        },
        "additionalProperties": false
      },
...
      "DeliveryDeleteBranchRequest": {
        "type": "object",
        "required": [
          "client_request_id",
          "expected_disposition_version",
          "expected_merge_operation_id",
          "expected_source_ref",
          "expected_source_oid",
          "target_branch",
          "target_head"
        ],
        "properties": {
          "client_request_id": {
            "type": "string",
            "format": "uuid",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "expected_disposition_version": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1
          },
          "expected_merge_operation_id": {
            "type": "string",
            "format": "uuid",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "expected_source_oid": {
            "type": "string",
            "maxLength": 64,
            "minLength": 40,
            "pattern": "^([0-9a-f]{40}|[0-9a-f]{64})$"
          },
          "expected_source_ref": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 12,
            "pattern": "^refs/heads/.{1,4085}$"
          },
          "target_branch": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 12,
            "pattern": "^refs/heads/.{1,4085}$"
          },
          "target_head": {
            "type": "string",
            "maxLength": 64,
            "minLength": 40,
            "pattern": "^([0-9a-f]{40}|[0-9a-f]{64})$"
          }
        },
        "additionalProperties": false
      },
...
      "DeliveryMergeRequest": {
        "type": "object",
        "required": [
          "client_request_id",
          "preflight_operation_id",
          "expected_operation_version",
          "expected_review_generation",
          "expected_workspace_fingerprint",
          "target_branch",
          "expected_target_head"
        ],
        "properties": {
          "client_request_id": {
            "type": "string",
            "format": "uuid",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "expected_operation_version": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1
          },
          "expected_review_generation": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 0
          },
          "expected_target_head": {
            "type": "string",
            "maxLength": 64,
            "minLength": 40,
            "pattern": "^([0-9a-f]{40}|[0-9a-f]{64})$"
          },
          "expected_workspace_fingerprint": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$"
          },
          "preflight_operation_id": {
            "type": "string",
            "format": "uuid",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "target_branch": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 12,
            "pattern": "^refs/heads/.{1,4085}$"
          }
        },
        "additionalProperties": false
      },
...
      "DeliveryOperationDto": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DeliveryMergeOperationEnvelopeDto"
          },
          {
            "$ref": "#/components/schemas/DeliveryCleanupOperationEnvelopeDto"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "cleanup": "#/components/schemas/DeliveryCleanupOperationEnvelopeDto",
            "merge": "#/components/schemas/DeliveryMergeOperationEnvelopeDto"
          }
        }
      },
...
      "DeliveryPreflightRequest": {
        "type": "object",
        "required": [
          "client_request_id",
          "target_branch",
          "expected_target_head"
        ],
        "properties": {
          "client_request_id": {
            "type": "string",
            "format": "uuid",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "expected_target_head": {
            "type": "string",
            "maxLength": 64,
            "minLength": 40,
            "pattern": "^([0-9a-f]{40}|[0-9a-f]{64})$"
          },
          "target_branch": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 12,
            "pattern": "^refs/heads/.{1,4085}$"
          }
        },
        "additionalProperties": false
      },
...
      "DeliveryReceiptDispositionDto": {
        "type": "string",
        "enum": [
          "created",
          "existing"
        ]
      },
      "DeliveryRemoveWorktreeRequest": {
        "type": "object",
        "required": [
          "client_request_id",
          "expected_disposition_version",
          "expected_merge_operation_id",
          "expected_source_ref",
          "expected_source_oid"
        ],
        "properties": {
          "client_request_id": {
            "type": "string",
            "format": "uuid",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "expected_disposition_version": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1
          },
          "expected_merge_operation_id": {
            "type": "string",
            "format": "uuid",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "expected_source_oid": {
            "type": "string",
            "maxLength": 64,
            "minLength": 40,
            "pattern": "^([0-9a-f]{40}|[0-9a-f]{64})$"
          },
          "expected_source_ref": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 12,
            "pattern": "^refs/heads/.{1,4085}$"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}